% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_time_restriciton}
\alias{apply_time_restriciton}
\title{Time filtering of data.table or sql-strings.}
\usage{
apply_time_restriciton(
  data,
  filter_colname,
  lower_limit,
  upper_limit,
  system_type
)
}
\arguments{
\item{data}{If system_type is a database, the sql-string goes here.
If system_type is 'csv', the data.table of this csv goes here.}

\item{filter_colname}{The name of the column to apply the time-filtering to.}

\item{lower_limit}{The posixct timestamp of the lower filtering boundary.}

\item{upper_limit}{The posixct timestamp of the upper filtering boundary.}

\item{system_type}{'postgres'/'oracle'/'csv'}
}
\value{
If system_type is a database, the new sql-string containing the
  temporal filtering will be returned ('order by' parts will be removed).
  If system_type is 'csv', the filtered data.table will be returned.
}
\description{
Internal function to filter the input data (or SQL) depending
  on provided time information. Sensitive to SQL dialects.
}
