% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_results.R
\name{atemp_pausi_results}
\alias{atemp_pausi_results}
\title{atemp_pausi_results helper function}
\usage{
atemp_pausi_results(rv, headless = FALSE)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}
}
\description{
Internal function to generate the results of the
  'Atemporal Plausibility' checks.
}
