% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date_restriction_requirements}
\alias{check_date_restriction_requirements}
\title{Checking the mdr integrity for time restrictions}
\usage{
check_date_restriction_requirements(
  mdr,
  system_names,
  restricting_date,
  logfile_dir,
  headless = TRUE,
  enable_stop = TRUE
)
}
\arguments{
\item{mdr}{The mdr as data.table}

\item{logfile_dir}{The absolute path to folder where the logfile
will be stored.}

\item{enable_stop}{(boolean, default = TRUE) If true (default) this function
will call `stop()` in case of the check fails. If `enable_stop = FALSE`
it will return `TRUE` if the check was successfull and `FALSE` if the
check failed. Use `enable_stop = FALSE` to avoid the need of a try/catch
block around this function.}
}
\value{
TRUE/FALSE: TRUE if the check was successfull and the given
  systems can be time filtered, FALSE if something went wrong and no time
  filtering is possible.
}
\description{
Internal function to check if for every input table there is
  one single (or empty) column where to apply the time restriction to.
  If the input is valid, it will just print a success-message, if the
  data is invalid, the function will call `stop()`.
}
