% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_file_current_runtime_id}
\alias{check_file_current_runtime_id}
\title{Is current runtime_id the one in this file?}
\usage{
check_file_current_runtime_id(path_with_file)
}
\arguments{
\item{path_with_file}{The path with the file to look at}
}
\description{
Helper function for the feedback function, especially
  the logfile function. Extracts the runtime_id from the
  logfile and compares it to the current runtime_id.
  If equal, return = TRUE.
}
