% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanPathName_}
\alias{cleanPathName_}
\title{cleanPathName_ helper function}
\usage{
cleanPathName_(pathname)
}
\arguments{
\item{pathname}{A character string. A pathname to be cleaned (to have a tailing slash).}
}
\description{
Internal function to clean paths to have a tailing slash
}
