% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanup_old_logfile}
\alias{cleanup_old_logfile}
\title{Archives the current logfile and creates a new blank one.}
\usage{
cleanup_old_logfile(logfile_dir)
}
\arguments{
\item{logfile_dir}{The absolute path to folder where the logfile
will be stored.}
}
\description{
This function is called once at the beginning of the
  runtime of the tool. It checks whether there is an old logfile
  and renames it (if existing) to "logfile_20xx-xx-xx-xxxxxx.log".
  Then a new, empty, logfile "logfile.log" is created.
}
