% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{completeness}
\alias{completeness}
\title{completeness helper function}
\usage{
completeness(results, headless = FALSE, logfile_dir)
}
\arguments{
\item{results}{A list object. The list should contain the results of
either 'rv$results_descriptive' or 'rv$results_plausibility_atemporal'.}
}
\description{
Internal function to perform missing analysis.
}
