% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{completeness_}
\alias{completeness_}
\title{completeness_ helper function}
\usage{
completeness_(results, headless = FALSE)
}
\arguments{
\item{results}{A list object. The list should contain the results of either 'rv$results_descriptive' or 'rv$results_plausibility_atemporal'.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is run only in the console (headless = TRUE) or on a GUI frontend (headless = FALSE).}
}
\description{
Internal function to perform missing analysis.
}
