% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr.R
\name{create_helper_vars}
\alias{create_helper_vars}
\title{create_helper_vars helper function}
\usage{
create_helper_vars(mdr, source_db, target_db)
}
\arguments{
\item{mdr}{A data.table object containing the MDR.}

\item{source_db}{A character string. The name of the source database.
This string must be conform with the corresponding config section
in the config.yml-file.}

\item{target_db}{A character string. The name of the target database.
This string must be conform with the corresponding config section
in the config.yml-file.}
}
\description{
Internal function to create necessary variables from the
  meta data repository (MDR).
}
\examples{
utils_path <- utils_path <- system.file(
  "demo_data/utilities",
  package = "DQAstats"
)
mdr_filename <- "mdr_example_data.csv"
mdr <- read_mdr(
  utils_path = utils_path,
  mdr_filename <- mdr_filename
)

source_system_name <- "exampleCSV_source"
target_system_name <- "exampleCSV_target"

create_helper_vars(
  mdr = mdr,
  source_db = source_system_name,
  target_db = target_system_name
)

}
