% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_markdown}
\alias{create_markdown}
\title{create_markdown helper function}
\usage{
create_markdown(rv = rv, utils, outdir = tempdir(), headless = FALSE)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}

\item{utils}{A character string. The path to the utils-folder,
containing the required app utilities like the MDR and the settings folder.}

\item{outdir}{A character string. The directory to store the resulting
PDF document. Default: \code{tempdir}.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}
}
\description{
Internal function to generate the final PDF report.
}
