% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{createMarkdown_}
\alias{createMarkdown_}
\title{createMarkdown_ helper function}
\usage{
createMarkdown_(utils, outdir = tempdir(), headless = FALSE)
}
\arguments{
\item{utils}{A character string. The path to the utils-folder, containing the requires app utilities. For a detailed description please visit \url{#TODO}.}

\item{outdir}{A character string. The directory to store the resulting PDF document. Default: \code{tempdir}.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is run only in the console (headless = TRUE) or on a GUI frontend (headless = FALSE).}
}
\description{
Internal function to generate the final PDF report.
}
