% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_results.R
\name{descriptive_results}
\alias{descriptive_results}
\title{descriptive_results helper function}
\usage{
descriptive_results(rv, headless = FALSE)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}
}
\description{
Internal function to generate the descriptive results.
}
