% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQA.R
\name{DQA}
\alias{DQA}
\title{Perform Data Quality Assessment of Electronic Health Records.}
\usage{
DQA(target_config, source_config, target_db, source_db, utils)
}
\arguments{
\item{target_config}{A character string. The path to the config.yml-file containing the target database configuration.}

\item{source_config}{A character string. The path to the config.yml-file containing the source database configuration.}

\item{target_db}{A character string. The name of the target database. This string must be conform with the corresponding config section in the config.yml-file.}

\item{source_db}{A character string. The name of the source database. This string must be conform with the corresponding config section in the config.yml-file.}

\item{utils}{A character string. The path to the utils-folder, containing the requires app utilities. For a detailed description please visit \url{#TODO}.}
}
\description{
This function performs a data quality assessment (DQA) of electronic health records (EHR).#'
}
\examples{
\dontrun{
DQA("type1_experimentaldata.csv", "type1_calibrationdata.csv", samplelocusname = "BRAF")
}

}
