% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{etl_checks}
\alias{etl_checks}
\title{etl_checks helper function}
\usage{
etl_checks(results)
}
\arguments{
\item{results}{A list object. The list should contain the results
'rv$results_descriptive'.}
}
\description{
Internal function to perform etl conformance checks.
}
