% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{etlChecks_}
\alias{etlChecks_}
\title{etlChecks_ helper function}
\usage{
etlChecks_(results)
}
\arguments{
\item{results}{A list object. The list should contain the results 'rv$results_descriptive'.}
}
\description{
Internal function to perform etl conformance checks.
}
