% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{export_affected_ids}
\alias{export_affected_ids}
\title{Export results to csv/zip file.}
\usage{
export_affected_ids(object, output_dir, rv)
}
\arguments{
\item{object}{The object to analyse.}

\item{output_dir}{The path to the output folder where all the results will
be stored (default: `paste0(tempdir(), "/output/")`).}

\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}
}
\description{
This function exports export_affected_ids in csv files that
  are added to a zip archive.
}
