% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{export_aggregated}
\alias{export_aggregated}
\title{Export results to csv/zip file.}
\usage{
export_aggregated(output_dir, rv)
}
\arguments{
\item{output_dir}{The path to the output folder where all the results will
be stored.}

\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}
}
\description{
This function exports aggregated results in csv files that
  are added to a zip archive.
}
