% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feedback_to_console}
\alias{feedback_to_console}
\title{Print to the console. Internal use.}
\usage{
feedback_to_console(print_this, type, findme, prefix, suffix)
}
\arguments{
\item{print_this}{(Optional, String)}

\item{type}{(Optional, String) E.g. "Warning", "Error. Default: "Info"}

\item{findme}{(Optional, String) String to find the message in the code.
E.g. 10-digit random hex from https://www.browserling.com/tools/random-hex
or https://onlinerandomtools.com/generate-random-hexadecimal-numbers}

\item{prefix}{Prefix (Optional, String) This is useful if
print_this is an array/list.
Each entry will then be new row with this prefix.}

\item{suffix}{Suffix (Optional, String). Same like prefix but at the
end of each line.}
}
\description{
Helper function for the feedback function to print
  stuff to the console. Everything will also be added to the logfile.
  Internal use. Use the robust 'feedback' function instead.
}
