% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feedback_to_logfile}
\alias{feedback_to_logfile}
\title{Add to the logfile. Internal use.}
\usage{
feedback_to_logfile(print_this, type, findme, prefix, suffix)
}
\arguments{
\item{input}{The input string to be added to the logfile.}
}
\description{
Helper function for the feedback function to add content
  to the logfile. Internal use.
  Use the robust 'feedback' function instead.
}
