% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feedback_to_logjs}
\alias{feedback_to_logjs}
\title{Feedback to the gui/browser-console with logjs. Internal use.}
\usage{
feedback_to_logjs(print_this, logfile_dir, headless)
}
\arguments{
\item{print_this}{(Optional, String)}

\item{logfile_dir}{The absolute path to folder where the logfile
will be stored.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}
}
\description{
Helper function for the feedback function to also show the
  messages to the gui/user via the browser console.
  Internal use. Use the robust 'feedback' function instead.
}
