% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feedback_to_ui}
\alias{feedback_to_ui}
\title{Feedback to the user with a modal. Internal use.}
\usage{
feedback_to_ui(print_this, type, logfile_dir, headless)
}
\arguments{
\item{print_this}{(Optional, String)}

\item{type}{(Optional, String) E.g. "Warning", "Error. Default: "Info"}

\item{logfile_dir}{The absolute path to folder where the logfile
will be stored.}
}
\description{
Helper function for the feedback function to show modals
  to the gui/user. Everything will also be added to the logfile.
  Internal use. Use the robust 'feedback' function instead.
}
