% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{firstup}
\alias{firstup}
\title{Converts the first letter of the input string to uppercase}
\usage{
firstup(x)
}
\arguments{
\item{x}{A character string. E.g. "hello world" will become "Hello world".}
}
\description{
Converts the first letter of the input string to uppercase
}
