% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap.R
\name{generate_datamap}
\alias{generate_datamap}
\title{generate_datamap helper function}
\usage{
generate_datamap(results, mdr, db, rv, headless = FALSE)
}
\arguments{
\item{results}{A list object. The list should contain the results
'rv$results_descriptive'.}

\item{mdr}{A data.table object containing the MDR.}

\item{db}{A character string. The name of the corresponding database.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}
}
\description{
Internal function to generate the dashboard data maps
}
