% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap.R
\name{generateDatamap_}
\alias{generateDatamap_}
\title{generateDatamap_ helper function}
\usage{
generateDatamap_(results, mdr, source_db, headless = FALSE)
}
\arguments{
\item{results}{A list object. The list should contain the results 'rv$results_descriptive'.}

\item{mdr}{A data.table object containing the MDR.}

\item{source_db}{A character string. The name of the source database. This string must be conform with the corresponding config section in the config.yml-file.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is run only in the console (headless = TRUE) or on a GUI frontend (headless = FALSE).}
}
\description{
Internal function to generate the dashboard data maps
}
