% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_config}
\alias{get_config}
\title{get_config helper function}
\usage{
get_config(config_file, config_key)
}
\arguments{
\item{config_file}{A character string. The path to the config.yml-file
containing the database configuration.}

\item{config_key}{A character string. The name of the corresponding
database. This string must be conform with the corresponding config
section in the config.yml-file.}
}
\description{
Internal function to read config files
}
