% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_new_runtime_id}
\alias{get_new_runtime_id}
\title{Creates a new random runtime_id with certain length}
\usage{
get_new_runtime_id(runtime_id_length)
}
\arguments{
\item{runtime_id_length}{The length of the resulting runtime_id}
}
\description{
Helper function for the feedback function, especially
  the logfile function. Creates a new random runtime_id
  with the given length.
}
