% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_runtime_id}
\alias{get_runtime_id}
\title{Returns the current runtime_id and stores it to rv$runtime_id}
\usage{
get_runtime_id(force = FALSE)
}
\arguments{
\item{force}{If true, a new runtime_id will be created.
If false (default) it depends wether there already is one or not.}
}
\description{
Helper function for the feedback function, especially
  the logfile function. If there is already a runtime_id, the current
  one will be returned. Otherwise a new one will be set,
  stored to rv$runtime_id and also be returned.
}
