% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{global_env_hack}
\alias{global_env_hack}
\title{global_env_hack}
\usage{
global_env_hack(key, val, pos = 1)
}
\arguments{
\item{key}{A character string. The name of the assigned variable}

\item{val}{An object. The object that will be assigned to 'key'.}

\item{pos}{An integer. The position of the environment (default: 1).}
}
\description{
Hack variable into global env (bypasses R CMD checks).
}
\seealso{
\url{http://adv-r.had.co.nz/Environments.html}
}
