% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloading.R
\name{load_csv}
\alias{load_csv}
\title{load_csv helper function}
\usage{
load_csv(rv, keys_to_test, headless = FALSE, system)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}

\item{keys_to_test}{A vector containing the names (keys) of
the variables to test.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}

\item{system}{The system object rv$system}
}
\description{
Internal function to load the source data
}
