% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_sqls}
\alias{load_sqls}
\title{load_sqls helper function}
\usage{
load_sqls(utils, db)
}
\arguments{
\item{utils}{A character string. The path to the utils-folder,
containing the required app utilities like the MDR and the settings folder.}

\item{db}{A character string. The name of the corresponding database.}
}
\description{
Internal function to load the SQL statements.
}
