% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloading.R
\name{load_target}
\alias{load_target}
\title{load_target helper function}
\usage{
load_target(rv, keys_to_test, headless = FALSE)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}

\item{keys_to_test}{A vector containing the names (keys) of
the variables to test.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}
}
\description{
Internal function to load the target data
}
