% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloading.R
\name{loadSource_}
\alias{loadSource_}
\title{loadSource_ helper function}
\usage{
loadSource_(rv, keys_to_test, headless = FALSE)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's 'reactive values'.}

\item{keys_to_test}{A vector containing the names (keys) of the variables to test.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is run only in the console (headless = TRUE) or on a GUI frontend (headless = FALSE).}
}
\description{
Internal function to load the source data
}
