% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadSQLs_}
\alias{loadSQLs_}
\title{loadSQLs_ helper function}
\usage{
loadSQLs_(utils, db)
}
\arguments{
\item{utils}{A character string. The path to the utils-folder, containing the requires app utilities. For a detailed description please visit \url{#TODO}.}

\item{db}{A character string. The name of the corresponding database.}
}
\description{
Internal function to load the SQL statements.
}
