% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logfile_assign_to_global}
\alias{logfile_assign_to_global}
\title{Stores logfile_dir to global environment}
\usage{
logfile_assign_to_global(logfile_dir_tmp)
}
\arguments{
\item{logfile_dir_tmp}{Path to the logfile folder}
}
\description{
Stores logfile_dir to global environment
}
