% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallel}
\alias{parallel}
\title{parallel helper function}
\usage{
parallel(parallel, logfile_dir, ncores)
}
\arguments{
\item{parallel}{A boolean. If TRUE, initializing a \code{future::plan()}
for running the code (default: FALSE).}

\item{logfile_dir}{The absolute path to folder where the logfile
will be stored default(\code{tempdir()}).}

\item{ncores}{A integer. The number of cores to use. Caution: you would
probably like to choose a low number when operating on large datasets.
Default: 2.}
}
\value{
No return value. Depending on the specified arguments, this function
enables a parallel backend for faster computations.
}
\description{
Internal function to initialize the parallel backend.
}
\examples{
parallel(parallel = FALSE, logfile_dir = tempdir(), ncores = 1)

}
