% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr.R
\name{read_mdr}
\alias{read_mdr}
\title{read_mdr helper function}
\usage{
read_mdr(utils)
}
\arguments{
\item{utils}{A character string. The path to the utils-folder,
containing the requires app utilities.
For a detailed description please visit \url{#TODO}.}
}
\description{
Internal function to read the meta data repository (MDR).
}
