% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr.R
\name{readMDR_}
\alias{readMDR_}
\title{readMDR_ helper function}
\usage{
readMDR_(utils)
}
\arguments{
\item{utils}{A character string. The path to the utils-folder, containing the requires app utilities. For a detailed description please visit \url{#TODO}.}
}
\description{
Internal function to read the meta data repository (MDR).
}
