% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{reduce_cat}
\alias{reduce_cat}
\title{reduce_cat helper function}
\usage{
reduce_cat(data, levellimit = 25)
}
\arguments{
\item{data}{A list object. The object `rv$results_descriptive`.}

\item{levellimit}{An integer value. The number of maximum
values to be displayed (default: 25).}
}
\description{
Internal function to reduce categorical variables
  to a maximum of values to be displayed.
}
