% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{test_db}
\alias{test_db}
\title{test_db helper function}
\usage{
test_db(settings, headless = FALSE, timeout = 30)
}
\arguments{
\item{settings}{A list object containing the database settings.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}

\item{timeout}{A timeout in sec. for the db-connection establishment.
Values below 2 seconds are not recommended.
Default is 30 seconds.}
}
\description{
Internal function to test and get the database connection of
  the target data system.
}
