% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{test_source_db}
\alias{test_source_db}
\title{test_source_db helper function}
\usage{
test_source_db(source_settings, source_db, headless = FALSE)
}
\arguments{
\item{source_settings}{A list object containing the database settings.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}
}
\description{
Internal function to test and get the database connection
  of the source data system.
}
