% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{test_target_db}
\alias{test_target_db}
\title{test_target_db helper function}
\usage{
test_target_db(target_settings, headless = FALSE)
}
\arguments{
\item{target_settings}{A list object containing the database settings.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}
}
\description{
Internal function to test and get the database connection of
  the target data system.
}
