% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{testSourceDB_}
\alias{testSourceDB_}
\title{testSourceDB_ helper function}
\usage{
testSourceDB_(source_settings, source_db, headless = FALSE)
}
\arguments{
\item{source_settings}{A list object containing the database settings.}

\item{source_db}{A character string. The name of the source database. This string must be conform with the corresponding config section in the config.yml-file.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is run only in the console (headless = TRUE) or on a GUI frontend (headless = FALSE).}
}
\description{
Internal function to test and get the database connection of the source data system.
}
