% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{testTargetDB_}
\alias{testTargetDB_}
\title{testTargetDB_ helper function}
\usage{
testTargetDB_(target_settings, headless = FALSE)
}
\arguments{
\item{target_settings}{A list object containing the database settings.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is run only in the console (headless = TRUE) or on a GUI frontend (headless = FALSE).}
}
\description{
Internal function to test and get the database connection of the target data system.
}
