% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_interval}
\alias{time_interval}
\title{time_interval helper function}
\usage{
time_interval(data)
}
\arguments{
\item{data}{The list object
'rv$results_descriptive$EpisodeOfCare_period_end'}
}
\description{
Internal function to get the time interval
}
