% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timeInterval_}
\alias{timeInterval_}
\title{timeInterval_ helper function}
\usage{
timeInterval_(data)
}
\arguments{
\item{data}{The list object 'rv$results_descriptive$EpisodeOfCare_period_end'}
}
\description{
Internal function to get the time interval
}
