% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_factors}
\alias{transform_factors}
\title{transform_factors helper function}
\usage{
transform_factors(vector, transformation)
}
\arguments{
\item{vector}{A vector containing the data that is to be transformed.}

\item{transformation}{A character string. The name of the (predefined)
transformation.}
}
\description{
Internal function to transform factors of the German §21-EHR
  billing data format
}
