% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformFactors_}
\alias{transformFactors_}
\title{transformFactors_ helper function}
\usage{
transformFactors_(vector, transformation)
}
\arguments{
\item{vector}{A vector containing the data that is to be transformed.}

\item{transformation}{A character string. The name of the (predefined) transformation.}
}
\description{
Internal function to transform factors of the German §21-EHR billing data format
}
