% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_results.R
\name{uniq_plausi_results}
\alias{uniq_plausi_results}
\title{uniq_plausi_results helper function}
\usage{
uniq_plausi_results(rv, uniq_vars, mdr, headless = FALSE)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}

\item{uniq_vars}{A data.table object. The object is created
by \code{create_helper_vars} from the data represented in
the metadata repository.}

\item{mdr}{A data.table object containing the MDR.}
}
\description{
Internal function to generate the results of
  the 'Uniqueness Plausibility' checks.
}
