% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibilityResults.R
\name{uniqPausiResults_}
\alias{uniqPausiResults_}
\title{uniqPausiResults_ helper function}
\usage{
uniqPausiResults_(rv, pl.uniq_vars, mdr, source_db, headless = FALSE)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's 'reactive values'.}

\item{pl.uniq_vars}{A data.table object. The object is created by \code{createHelperVars_} from the data represented in the metadata repository.}

\item{mdr}{A data.table object containing the MDR.}

\item{source_db}{A character string. The name of the source database. This string must be conform with the corresponding config section in the config.yml-file.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is run only in the console (headless = TRUE) or on a GUI frontend (headless = FALSE).}
}
\description{
Internal function to generate the results of the 'Uniqueness Plausibility' checks.
}
