% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformance.R
\name{value_conformance}
\alias{value_conformance}
\title{value_conformance helper function}
\usage{
value_conformance(rv, results, scope, headless = FALSE, logfile_dir)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}

\item{results}{A list object. The list should contain the results of
either 'rv$results_descriptive' or 'rv$results_plausibility_atemporal'.}

\item{scope}{A character. Either "plausibility" or "descriptive".}
}
\description{
Internal function to perform value conformance checks.
}
