% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformance.R
\name{value_conformance_checks}
\alias{value_conformance_checks}
\title{value_conformance_checks helper function}
\usage{
value_conformance_checks(results)
}
\arguments{
\item{results}{A list object. The list should contain the results of
the function \code{value_conformance}.}
}
\description{
Internal function to perform value conformance checks.
}
