% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformance.R
\name{valueConformance_}
\alias{valueConformance_}
\title{valueConformance_ helper function}
\usage{
valueConformance_(results, headless = FALSE)
}
\arguments{
\item{results}{A list object. The list should contain the results of either 'rv$results_descriptive' or 'rv$results_plausibility_atemporal'.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is run only in the console (headless = TRUE) or on a GUI frontend (headless = FALSE).}
}
\description{
Internal function to perform value conformance checks.
}
