% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformance.R
\name{valueConformanceChecks_}
\alias{valueConformanceChecks_}
\title{valueConformanceChecks_ helper function}
\usage{
valueConformanceChecks_(results)
}
\arguments{
\item{results}{A list object. The list should contain the results of the function \code{valueConformance_}.}
}
\description{
Internal function to perform value conformance checks.
}
