% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date_restriction_requirements}
\alias{check_date_restriction_requirements}
\title{Checking the mdr integrity for time restrictions}
\usage{
check_date_restriction_requirements(mdr, restricting_date, logfile_dir)
}
\arguments{
\item{mdr}{The mdr as data.table}

\item{logfile_dir}{The absolute path to folder where the logfile
will be stored.}

\item{restriction_date}{(list) If `restriction_date$use_it == FALSE`,
the result will always be true since it doesn't matter if the restriction
parameters are valid, because we don't use them.}
}
\description{
Internal function to check if for every input table there is
  one single (or empty) column where to apply the time restriction to.
  If the input is valid, it will just print a success-message, if the
  data is invalid, the function will call `stop()`.
}
