% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{close_all_connections}
\alias{close_all_connections}
\title{Cleanup function to unset/close all open connections}
\usage{
close_all_connections(logfile_dir, headless)
}
\arguments{
\item{logfile_dir}{The absolute path to folder where the logfile
will be stored.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}
}
\description{
This function is meant to be called at the end of a
  run of the app. It will close all open connections to files.
}
