% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr.R
\name{create_helper_vars}
\alias{create_helper_vars}
\title{create_helper_vars helper function}
\usage{
create_helper_vars(mdr, target_db, source_db)
}
\arguments{
\item{mdr}{A data.table object containing the MDR.}
}
\description{
Internal function to create necessar variables from the
  meta data repository (MDR).
}
