% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr.R
\name{create_helper_vars}
\alias{create_helper_vars}
\title{create_helper_vars helper function}
\usage{
create_helper_vars(mdr, target_db, source_db)
}
\arguments{
\item{mdr}{A data.table object containing the MDR.}

\item{target_db}{A character string. The name of the target database.
This string must be conform with the corresponding config section
in the config.yml-file.}

\item{source_db}{A character string. The name of the source database.
This string must be conform with the corresponding config section
in the config.yml-file.}
}
\description{
Internal function to create necessary variables from the
  meta data repository (MDR).
}
