% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_results.R
\name{descriptive_results}
\alias{descriptive_results}
\title{descriptive_results helper function}
\usage{
descriptive_results(rv, headless = FALSE)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}
}
\description{
Internal function to generate the descriptive results.
}
