% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiveResults.R
\name{descriptiveResults_}
\alias{descriptiveResults_}
\title{descriptiveResults_ helper function}
\usage{
descriptiveResults_(rv, headless = FALSE)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's 'reactive values'.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is run only in the console (headless = TRUE) or on a GUI frontend (headless = FALSE).}
}
\description{
Internal function to generate the descriptive results.
}
