% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQA.R
\name{dqa}
\alias{dqa}
\title{Perform Data Quality Assessment of Electronic Health Records.}
\usage{
dqa(
  source_system_name,
  target_system_name,
  config_file,
  utils_path,
  mdr_filename = "mdr.csv",
  output_dir = "./output/"
)
}
\arguments{
\item{source_system_name}{A character string. The name of the
source-system, e.g. "P21" or "i2b2". This name must be identical and
unique to one entry in the settings-yml file.}

\item{target_system_name}{Optional. A character string or null.
The name of the target-system, e.g. "P21" or "i2b2".
This name must be identical and unique to one entry in the
config-yml file or null. If the argument is empty, the source will
be processed as standalone on its own.}

\item{config_file}{The config.yml-file containig all the information
needed to access the source (and optional the target) system(s).}

\item{utils_path}{A character string. The path to the utils-folder,
containing the required app utilities like the MDR and the settings folder.}

\item{mdr_filename}{A character string.
The filename of the MDR e.g. "mdr_example_data.csv"
For a detailed description please visit \url{#TODO}.}

\item{output_dir}{The path to the output folder where all the results will
be stored.}
}
\description{
This function performs a data quality assessment (DQA)
  of electronic health records (EHR).#'
}
\examples{
\dontrun{
DQA("type1_experimentaldata.csv", "type1_calibrationdata.csv",
samplelocusname = "BRAF")
}

}
