% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feedback_to_logjs}
\alias{feedback_to_logjs}
\title{Feedback to the gui/browser-console with logjs. Internal use.}
\usage{
feedback_to_logjs(print_this)
}
\arguments{
\item{print_this}{(Optional, String)}
}
\description{
Helper function for the feedback function to also show the
  messages to the gui/user via the browser console.
  Internal use. Use the robust 'feedback' function instead.
}
