% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibilities.R
\name{get_atemp_plausis}
\alias{get_atemp_plausis}
\title{get_atemp_plausis helper function}
\usage{
get_atemp_plausis(rv, atemp_vars, mdr, headless = FALSE)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}

\item{atemp_vars}{A data.table object. The object is created
by \code{create_helper_vars} from the data represented in the
metadata repository.}

\item{mdr}{A data.table object containing the MDR.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}
}
\description{
Internal function to generate raw data for the
  'Atemporal Plausibility' checks.
}
