% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_restrictions.R
\name{get_restricting_date_info}
\alias{get_restricting_date_info}
\title{Get a formatted string containing start and end time of the
date restriction applied to the data.}
\usage{
get_restricting_date_info(
  restricting_date,
  lang = "en",
  date = TRUE,
  time = TRUE
)
}
\arguments{
\item{restricting_date}{The list applied from rv$restricting_date}

\item{lang}{Language of the result. "de"/"en" (en = default).
If language is not yet implemented, "en" is used.}

\item{date}{Should the date be included in the result string?}

\item{time}{Should the time be included in the result string?}
}
\value{
String containing start and end date obtained from the list of
\code{restricting_date}.
}
\description{
See title.
}
