% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloading.R
\name{load_database}
\alias{load_database}
\title{load_database helper function}
\usage{
load_database(
  rv,
  sql_statements,
  db_con,
  keys_to_test,
  db_name,
  headless = FALSE
)
}
\arguments{
\item{rv}{A list object. Internal list simulating Shiny's
'reactive values'.}

\item{sql_statements}{The SQL-Statement-object}

\item{db_con}{The connection-socket}

\item{keys_to_test}{A vector containing the names (keys) of
the variables to test.}

\item{db_name}{The database name}
}
\description{
Internal function to load the target data
}
