% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallel}
\alias{parallel}
\title{parallel helper function}
\usage{
parallel(parallel, logfile_dir, ncores)
}
\arguments{
\item{parallel}{A boolean. If TRUE (the default value), initializing
`future::plan("multiprocess")` before running the code.}

\item{logfile_dir}{The absolute path to folder where the logfile
will be stored.}

\item{ncores}{A integer. The number of cores to use. Caution: you would
probably like to choose a low number when operating on large datasets.
Default: 2}
}
\description{
Internal function to initialize the parallel backend.
}
