% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{test_csv}
\alias{test_csv}
\title{test_csv helper function}
\usage{
test_csv(settings, source_db, mdr, headless = FALSE, logfile_dir)
}
\arguments{
\item{settings}{A list object containing the database settings.}

\item{source_db}{A character string. The name of the source database.
This string must be conform with the corresponding config section
in the config.yml-file.}

\item{mdr}{A data.table object containing the MDR.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}

\item{logfile_dir}{The absolute path to folder where the logfile
will be stored.}
}
\description{
Internal function to test and get the database connection
  of the source data system.
}
