% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformance.R
\name{value_conformance}
\alias{value_conformance}
\title{value_conformance helper function}
\usage{
value_conformance(results, headless = FALSE)
}
\arguments{
\item{results}{A list object. The list should contain the results of
either 'rv$results_descriptive' or 'rv$results_plausibility_atemporal'.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}
}
\description{
Internal function to perform value conformance checks.
}
